seporate function of create pagination to utils
